Re-factor Copy component.
